move UserCard to widget, add cards to storybook
